+++ backtester/__init__.py
@@ -1,2 +1,2 @@
-"""backtester service v0.2.0"""
-__version__ = "0.2.0"
+"""backtester service v0.3.0 (2025-08-18)"""
+__version__ = "0.3.0"